requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
aiohappyeyeballs=2.4.0=pypi_0
aiohttp=3.10.5=pypi_0
aiosignal=1.3.1=pypi_0
alembic=1.13.2=pypi_0
altair=5.4.1=pypi_0
annotated-types=0.7.0=pypi_0
anyio=4.4.0=pypi_0
asgiref=3.8.1=pypi_0
async-timeout=4.0.3=pypi_0
attrs=24.2.0=pypi_0
backoff=2.2.1=pypi_0
bcrypt=4.2.0=pypi_0
blinker=1.8.2=pypi_0
build=1.2.2=pypi_0
bzip2=1.0.8=h5eee18b_6
ca-certificates=2024.7.2=h06a4308_0
cachetools=5.5.0=pypi_0
certifi=2024.8.30=pypi_0
charset-normalizer=3.3.2=pypi_0
chroma-hnswlib=0.7.6=pypi_0
chromadb=0.5.5=pypi_0
click=8.1.7=pypi_0
coloredlogs=15.0.1=pypi_0
colorlog=6.8.2=pypi_0
datasets=2.14.7=pypi_0
debugpy=1.8.5=pypi_0
deprecated=1.2.14=pypi_0
dill=0.3.7=pypi_0
distro=1.9.0=pypi_0
docstring-parser=0.16=pypi_0
dspy=0.1.5=pypi_0
dspy-ai=2.4.5=pypi_0
exceptiongroup=1.2.2=pypi_0
fastapi=0.114.2=pypi_0
fastembed=0.3.6=pypi_0
filelock=3.16.0=pypi_0
flatbuffers=24.3.25=pypi_0
frozenlist=1.4.1=pypi_0
fsspec=2023.10.0=pypi_0
gitdb=4.0.11=pypi_0
gitpython=3.1.43=pypi_0
google-auth=2.34.0=pypi_0
googleapis-common-protos=1.65.0=pypi_0
greenlet=3.1.0=pypi_0
grpcio=1.66.1=pypi_0
grpcio-tools=1.66.1=pypi_0
h11=0.14.0=pypi_0
h2=4.1.0=pypi_0
hpack=4.0.0=pypi_0
httpcore=1.0.5=pypi_0
httptools=0.6.1=pypi_0
httpx=0.27.2=pypi_0
huggingface-hub=0.24.7=pypi_0
humanfriendly=10.0=pypi_0
hyperframe=6.0.1=pypi_0
idna=3.10=pypi_0
importlib-metadata=8.4.0=pypi_0
importlib-resources=6.4.5=pypi_0
instructor=1.4.2=pypi_0
jinja2=3.1.4=pypi_0
jiter=0.5.0=pypi_0
joblib=1.3.2=pypi_0
jsonschema=4.23.0=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
kubernetes=30.1.0=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
loguru=0.7.2=pypi_0
mako=1.3.5=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.5=pypi_0
mdurl=0.1.2=pypi_0
mmh3=4.1.0=pypi_0
monotonic=1.6=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.1.0=pypi_0
multiprocess=0.70.15=pypi_0
narwhals=1.8.2=pypi_0
ncurses=6.4=h6a678d5_0
numpy=1.26.4=pypi_0
oauthlib=3.2.2=pypi_0
onnx=1.16.2=pypi_0
onnxruntime=1.19.2=pypi_0
openai=1.45.0=pypi_0
openssl=3.0.15=h5eee18b_0
opentelemetry-api=1.27.0=pypi_0
opentelemetry-exporter-otlp-proto-common=1.27.0=pypi_0
opentelemetry-exporter-otlp-proto-grpc=1.27.0=pypi_0
opentelemetry-instrumentation=0.48b0=pypi_0
opentelemetry-instrumentation-asgi=0.48b0=pypi_0
opentelemetry-instrumentation-fastapi=0.48b0=pypi_0
opentelemetry-proto=1.27.0=pypi_0
opentelemetry-sdk=1.27.0=pypi_0
opentelemetry-semantic-conventions=0.48b0=pypi_0
opentelemetry-util-http=0.48b0=pypi_0
optuna=4.0.0=pypi_0
orjson=3.10.7=pypi_0
overrides=7.7.0=pypi_0
packaging=24.1=pypi_0
pandas=2.2.2=pypi_0
pillow=10.4.0=pypi_0
pip=24.2=py310h06a4308_0
portalocker=2.10.1=pypi_0
posthog=3.6.5=pypi_0
protobuf=5.28.1=pypi_0
pyarrow=17.0.0=pypi_0
pyarrow-hotfix=0.6=pypi_0
pyasn1=0.6.1=pypi_0
pyasn1-modules=0.4.1=pypi_0
pydantic=2.9.2=pypi_0
pydantic-core=2.23.4=pypi_0
pydeck=0.9.1=pypi_0
pygments=2.18.0=pypi_0
pypika=0.48.9=pypi_0
pyproject-hooks=1.1.0=pypi_0
pystemmer=2.2.0.1=pypi_0
python=3.10.14=h955ad1f_1
python-dateutil=2.9.0.post0=pypi_0
python-dotenv=1.0.1=pypi_0
pytz=2024.2=pypi_0
pyyaml=6.0.2=pypi_0
qdrant-client=1.11.1=pypi_0
readline=8.2=h5eee18b_0
referencing=0.35.1=pypi_0
regex=2024.9.11=pypi_0
requests=2.32.3=pypi_0
requests-oauthlib=2.0.0=pypi_0
rich=13.8.1=pypi_0
rpds-py=0.20.0=pypi_0
rsa=4.9=pypi_0
setuptools=72.1.0=py310h06a4308_0
shellingham=1.5.4=pypi_0
six=1.16.0=pypi_0
smmap=5.0.1=pypi_0
sniffio=1.3.1=pypi_0
snowballstemmer=2.2.0=pypi_0
sqlalchemy=2.0.34=pypi_0
sqlite=3.45.3=h5eee18b_0
starlette=0.38.5=pypi_0
streamlit=1.38.0=pypi_0
sympy=1.13.2=pypi_0
tenacity=8.5.0=pypi_0
tk=8.6.14=h39e8969_0
tokenizers=0.20.0=pypi_0
toml=0.10.2=pypi_0
tomli=2.0.1=pypi_0
tornado=6.4.1=pypi_0
tqdm=4.66.5=pypi_0
typer=0.12.5=pypi_0
typing-extensions=4.12.2=pypi_0
tzdata=2024.1=pypi_0
ujson=5.10.0=pypi_0
urllib3=2.2.3=pypi_0
uvicorn=0.30.6=pypi_0
uvloop=0.20.0=pypi_0
watchdog=4.0.2=pypi_0
watchfiles=0.24.0=pypi_0
websocket-client=1.8.0=pypi_0
websockets=13.0.1=pypi_0
wheel=0.44.0=py310h06a4308_0
wrapt=1.16.0=pypi_0
xxhash=3.5.0=pypi_0
xz=5.4.6=h5eee18b_1
yarl=1.11.1=pypi_0
zipp=3.20.2=pypi_0
zlib=1.2.13=h5eee18b_1
